+++ src/main/java/com/mdrsolutions/thymeleaf/thymeleafsupport/thymeleaf/expression/DollarSignInsertHandler.java
@@ -1,0 +1,9 @@
+package com.mdrsolutions.thymeleaf.thymeleafsupport.thymeleaf.expression; // Adjust package as needed
+
+public class DollarSignInsertHandler extends ThymeleafExpressionInsertHandler {
+
+    @Override
+    protected String getExpressionPrefix() {
+        return "${";
+    }
+}